Remove annoying grey/block line in watchface menu logo
The logo displayed in the watchface selection menu had a grey line (black in
the BW resource) at the right side, in the last column, which displayed
annoyingly in the menu. This fixes the issue, reworks the dithering of the
images, and renames the image file names so they do not contain "bw" in the
name (even the color resource had a "bw").